+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - added country encoded value for all countries (#2353)
 - removed spatial rules, replaced by country rules and custom areas (#2353)
+- removed api module and moved it into web-api, no more Jackson MixIns (#2372)
 - flag encoders are no longer versioned (#2355)
 - JSON route response contains now bbox if start and end are identical
 - renamed PriorityCode enums: AVOID_IF_POSSIBLE -> SLIGHT_AVOID, REACH_DEST -> AVOID, AVOID_AT_ALL_COSTS -> AVOID_MORE, WORST -> BAD